refactor and loop lesson creation
