+++ scripts/commands/create/lesson.js
@@ -1,13 +1,9 @@
-const { selectTutorial } = require('./utils.js')
-
-const fs = require('fs')
-const promisify = require('util').promisify
+const { selectTutorial, createLesson, logLessons } = require('./utils.js')
 
 const inquirer = require('inquirer')
 const log = require('npmlog')
 
 const run = require('../../modules/run')
-// const tutorials = require('../../../src/static/tutorials.json')
 
 // customize log styling
 log.addLevel('info', 2000, { fg: 'blue', bold: true }, '🧙‍♂️ ProtoWizard')
@@ -17,81 +13,19 @@ log.info("Let's add lessons to your tutorial.")
 async function command (options) {
   const { tutorial, tutorialId, lessons } = await selectTutorial('lesson')
 
-  // figure out next lesson lessonNumber
-  let lessonNumber = '01'
-  if (lessons.length > 0) {
-    lessonNumber = (lessons.map(lesson => lesson.id).sort().reverse()[0] + 1).toString().padStart(2, 0)
-  }
-
   // print existing lessons if present
   if (lessons.length === 0) {
     log.info("Let's create your first lesson!")
   } else {
     log.info("Here are the lessons you've created so far:")
-    console.group()
-    lessons.forEach(lesson => {
-      log.info(`${lesson.id} - ${lesson.title} (${lesson.type})`)
-    })
-    console.groupEnd()
+    logLessons(lessons)
     log.info("Let's create your next lesson!")
   }
 
-  const lessonResponses = await inquirer
-    .prompt([
-      {
-        type: 'input',
-        name: 'title',
-        message: "What's the title of this lesson?"
-      },
-      {
-        type: 'list',
-        name: 'type',
-        message: "What's the format of this lesson?",
-        choices: [
-          {
-            name: 'Text only',
-            value: 'text'
-          },
-          {
-            name: 'Multiple-choice quiz',
-            value: 'multiple-choice'
-          },
-          {
-            name: 'Code challenge',
-            value: 'code'
-          },
-          {
-            name: 'Code challenge with file upload',
-            value: 'file-upload'
-          }
-        ]
-      }
-    ])
+  // runs until you say you don't want more lessons
+  await createLesson(tutorial, tutorialId, lessons)
 
-  // create lesson
-  let newFileDetails = [`${lessonNumber}.md (for writing the text of your lesson)`]
-  await promisify(fs.copyFile)('src/tutorials/boilerplates/boilerplate.md', `src/tutorials/${tutorialId}-${tutorial.url}/${lessonNumber}.md`)
-  let markdown = await promisify(fs.readFile)(`src/tutorials/${tutorialId}-${tutorial.url}/${lessonNumber}.md`, 'utf8')
-  let newMarkdown = markdown.replace(`title: "Lesson title"`, `title: "${lessonResponses.title}"`)
-  if (lessonResponses.type !== 'text') {
-    newMarkdown = newMarkdown.replace(`type: "text"`, `type: "${lessonResponses.type}"`)
-    await promisify(fs.copyFile)(`src/tutorials/boilerplates/boilerplate-${lessonResponses.type}.js`, `src/tutorials/${tutorialId}-${tutorial.url}/${lessonNumber}.js`)
-    if (lessonResponses.type !== 'multiple-choice') {
-      await promisify(fs.copyFile)(`src/tutorials/boilerplates/boilerplate-challenge.md`, `src/tutorials/${tutorialId}-${tutorial.url}/${lessonNumber}-challenge.md`)
-      newFileDetails.push(`${lessonNumber}-challenge.md (for describing your code challenge)`)
-      newFileDetails.push(`${lessonNumber}.js (for building your code challenge)`)
-    } else {
-      newFileDetails.push(`${lessonNumber}.js (for building your multiple-choice quiz)`)
-    }
-  }
-  await promisify(fs.writeFile)(`src/tutorials/${tutorialId}-${tutorial.url}/${lessonNumber}.md`, newMarkdown)
-
-  // log success
-  log.info("Tada! We've created the following files that you'll need for this lesson:")
-  console.group()
-  newFileDetails.forEach(file => log.info(file))
-  console.groupEnd()
-  log.info(`Preview your tutorial by running \`npm start\` and visiting: http://localhost:3000/#/${tutorial.url}/${lessonNumber}`)
+  log.info(`We'll poke you here to create resources if you don't have any yet.`)
 }
 
 run(command)
